# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, data, file, plugin, mcp, skill, plus the sensitive
    config:edit_admin and chat:admin sub-scopes. The config scope also grants bot
    and provider access, but no parent scope grants a sensitive sub-scope.
...
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login to the dashboard API
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "401":
          $ref: "#/components/responses/Error"
...
  /api/v1/auth/setup:
    post:
      tags: [Auth]
      summary: Complete first-run dashboard account setup
      operationId: setupAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupAuthRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/auth/totp/setup:
    post:
      tags: [Auth]
      summary: Start or refresh TOTP setup
      operationId: setupTotp
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpSetupRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /api/v1/auth/account:
    patch:
      tags: [Auth]
      summary: Update dashboard account credentials
      operationId: updateAuthAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
    Error:
      description: Standard AstrBot error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
    ErrorEnvelope:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          const: error
        message:
          type: string
        data: {}
      additionalProperties: false
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        code:
          type: string
          description: TOTP code or recovery code when two-factor authentication is required.
        trust_device_flag:
          type: boolean
      additionalProperties: false
    SetupAuthRequest:
      type: object
      required: [username, password, confirm_password]
      properties:
        username:
          type: string
        password:
          type: string
        confirm_password:
          type: string
      additionalProperties: false
    UpdateAccountRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
        new_password:
          type: string
        confirm_password:
          type: string
        new_username:
          type: string
      additionalProperties: false
    TotpSetupRequest:
      type: object
      properties:
        code:
          type: string
        secret:
          type: string
      additionalProperties: false
...